# backend/internal/oapi/models.gen.go
// Package oapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package oapi

import (
	"time"
)

// Defines values for StatisticsAPIGetParamsPeriod.
const (
	Day   StatisticsAPIGetParamsPeriod = "day"
	Month StatisticsAPIGetParamsPeriod = "month"
	Week  StatisticsAPIGetParamsPeriod = "week"
)

// Error エラーレスポンス
type Error struct {
	Message string `json:"message"`
}

// Pager レスポンスのページネーション情報。
// ページに含まれる対象の数は以下で求められる。
//
// ```
// min(size, totalCount - size * (page - 1))
// ```
type Pager struct {
	// Page ページ番号
	Page int32 `json:"page"`

	// Size 1ページあたりの最大取得件数
	Size int32 `json:"size"`

	// TotalCount ページに含まれているものも含まれていないものも合わせた対象の全数
	TotalCount int32 `json:"totalCount"`
}

// PeriodStatistics 期間統計
type PeriodStatistics struct {
	// Items 統計アイテム一覧
	Items []StatisticsItem `json:"items"`

	// Total 合計時間(秒)
	Total int64 `json:"total"`
}

// StatisticsItem 統計アイテム
type StatisticsItem struct {
	// Duration 合計時間(秒)
	Duration int64 `json:"duration"`

	// Percentage 割合(%)
	Percentage float64 `json:"percentage"`

	// Title タイトル
	Title string `json:"title"`
}

// Video 録画ビデオ
type Video struct {
	// AudioActivityRatio 音声アクティビティ率(%)
	AudioActivityRatio float64 `json:"audioActivityRatio"`

	// CreatedAt 作成日時
	CreatedAt *time.Time `json:"createdAt,omitempty"`

	// Filename ファイル名
	Filename string `json:"filename"`

	// FinishedAt 録画終了時刻(UTC)
	FinishedAt time.Time `json:"finishedAt"`

	// Id ビデオID
	Id *int64 `json:"id,omitempty"`

	// StartedAt 録画開始時刻(UTC)
	StartedAt time.Time `json:"startedAt"`

	// Title タイトル
	Title string `json:"title"`

	// UpdatedAt 更新日時
	UpdatedAt *time.Time `json:"updatedAt,omitempty"`
}

// VideoCreate ビデオ作成リクエスト
type VideoCreate struct {
	// AudioActivityRatio 音声アクティビティ率(%)
	AudioActivityRatio float64   `json:"audioActivityRatio"`
	Filename           string    `json:"filename"`
	FinishedAt         time.Time `json:"finishedAt"`
	StartedAt          time.Time `json:"startedAt"`
	Title              string    `json:"title"`
}

// VideoPage ビデオページレスポンス
type VideoPage struct {
	Data []Video `json:"data"`

	// Pager レスポンスのページネーション情報。
	// ページに含まれる対象の数は以下で求められる。
	//
	// ```
	// min(size, totalCount - size * (page - 1))
	// ```
	Pager Pager `json:"pager"`
}

// VideoUpdate ビデオ更新リクエスト
type VideoUpdate struct {
	// AudioActivityRatio 音声アクティビティ率(%)
	AudioActivityRatio *float64   `json:"audioActivityRatio,omitempty"`
	Filename           *string    `json:"filename,omitempty"`
	FinishedAt         *time.Time `json:"finishedAt,omitempty"`
	StartedAt          *time.Time `json:"startedAt,omitempty"`
	Title              *string    `json:"title,omitempty"`
}

// StatisticsAPIGetParams defines parameters for StatisticsAPIGet.
type StatisticsAPIGetParams struct {
	// Period 期間
	Period StatisticsAPIGetParamsPeriod `form:"period" json:"period"`

	// Limit 取得件数
	Limit *int32 `form:"limit,omitempty" json:"limit,omitempty"`

	// BaseDate 基準日（YYYY-MM-DD形式）
	BaseDate *string `form:"baseDate,omitempty" json:"baseDate,omitempty"`

	// Timezone タイムゾーン（IANA timezone database形式、例：Asia/Tokyo）。期間の境界を計算するので必要
	Timezone *string `form:"timezone,omitempty" json:"timezone,omitempty"`
}

// StatisticsAPIGetParamsPeriod defines parameters for StatisticsAPIGet.
type StatisticsAPIGetParamsPeriod string

// VideosListParams defines parameters for VideosList.
type VideosListParams struct {
	// Page ページ番号
	Page *int32 `form:"page,omitempty" json:"page,omitempty"`

	// Size 1ページあたりの最大取得件数
	Size *int32 `form:"size,omitempty" json:"size,omitempty"`

	// StartedAtFrom 開始時刻の範囲指定（from）。この時刻以降に開始した動画を取得する
	StartedAtFrom *time.Time `form:"startedAtFrom,omitempty" json:"startedAtFrom,omitempty"`

	// StartedAtTo 開始時刻の範囲指定（to）。この時刻以前に開始した動画を取得する
	StartedAtTo *time.Time `form:"startedAtTo,omitempty" json:"startedAtTo,omitempty"`
}

// VideosCreateJSONRequestBody defines body for VideosCreate for application/json ContentType.
type VideosCreateJSONRequestBody = VideoCreate

// VideosUpdateJSONRequestBody defines body for VideosUpdate for application/json ContentType.
type VideosUpdateJSONRequestBody = VideoUpdate