# oas/openapi.yml
openapi: 3.0.0
info:
  title: TV API
  version: 0.0.0
tags: []
paths:
  /api/v1/statistics:
    get:
      operationId: StatisticsAPI_get
      description: 統計取得
      parameters:
        - name: period
          in: query
          required: true
          description: 期間
          schema:
            type: string
            enum:
              - day
              - week
              - month
          explode: false
        - name: limit
          in: query
          required: false
          description: 取得件数
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 10
          explode: false
        - name: baseDate
          in: query
          required: false
          description: 基準日（YYYY-MM-DD形式）
          schema:
            type: string
          explode: false
        - name: timezone
          in: query
          required: false
          description: タイムゾーン（IANA timezone database形式、例：Asia/Tokyo）。期間の境界を計算するので必要
          schema:
            type: string
            default: UTC
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodStatistics'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/videos:
    get:
      operationId: Videos_list
      description: ビデオ一覧取得
      parameters:
        - name: page
          in: query
          required: false
          description: ページ番号
          schema:
            type: integer
            format: int32
            minimum: 1
            default: 1
          explode: false
        - name: size
          in: query
          required: false
          description: 1ページあたりの最大取得件数
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 30
          explode: false
        - name: startedAtFrom
          in: query
          required: false
          description: 開始時刻の範囲指定（from）。この時刻以降に開始した動画を取得する
          schema:
            type: string
            format: date-time
          explode: false
        - name: startedAtTo
          in: query
          required: false
          description: 開始時刻の範囲指定（to）。この時刻以前に開始した動画を取得する
          schema:
            type: string
            format: date-time
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPage'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: Videos_create
      description: ビデオ作成
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoCreate'
  /api/v1/videos/{id}:
    get:
      operationId: Videos_get
      description: ビデオ詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: Videos_update
      description: ビデオ更新
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoUpdate'
    delete:
      operationId: Videos_delete
      description: ビデオ削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/videos/{id}/file:
    get:
      operationId: Videos_file
      description: ビデオファイル配信
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The request has succeeded.
          content:
            video/webm:
              schema:
                type: string
                format: binary
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/videos/{id}/thumbnail:
    get:
      operationId: Videos_thumbnail
      description: サムネイル画像取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The request has succeeded.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      description: エラーレスポンス
    Pager:
      type: object
      required:
        - page
        - size
        - totalCount
      properties:
        page:
          type: integer
          format: int32
          minimum: 1
          description: ページ番号
        size:
          type: integer
          format: int32
          minimum: 0
          maximum: 100
          description: 1ページあたりの最大取得件数
        totalCount:
          type: integer
          format: int32
          minimum: 0
          description: ページに含まれているものも含まれていないものも合わせた対象の全数
      description: |-
        レスポンスのページネーション情報。
        ページに含まれる対象の数は以下で求められる。

        ```
        min(size, totalCount - size * (page - 1))
        ```
    PeriodStatistics:
      type: object
      required:
        - items
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StatisticsItem'
          description: 統計アイテム一覧
        total:
          type: integer
          format: int64
          description: 合計時間(秒)
      description: 期間統計
    StatisticsItem:
      type: object
      required:
        - title
        - duration
        - percentage
      properties:
        title:
          type: string
          description: タイトル
        duration:
          type: integer
          format: int64
          description: 合計時間(秒)
        percentage:
          type: number
          format: double
          description: 割合(%)
      description: 統計アイテム
    Video:
      type: object
      required:
        - id
        - title
        - filename
        - startedAt
        - finishedAt
        - audioActivityRatio
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: ビデオID
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: タイトル
        filename:
          type: string
          minLength: 1
          maxLength: 255
          description: ファイル名
        startedAt:
          type: string
          format: date-time
          description: 録画開始時刻(UTC)
        finishedAt:
          type: string
          format: date-time
          description: 録画終了時刻(UTC)
        audioActivityRatio:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: 音声アクティビティ率(%)
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          readOnly: true
      description: 録画ビデオ
    VideoCreate:
      type: object
      required:
        - title
        - filename
        - startedAt
        - finishedAt
        - audioActivityRatio
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        filename:
          type: string
          minLength: 1
          maxLength: 255
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        audioActivityRatio:
          type: number
          format: double
          minimum: 0
          maximum: 100
          description: 音声アクティビティ率(%)
      description: ビデオ作成リクエスト
    VideoPage:
      type: object
      required:
        - pager
        - data
      properties:
        pager:
          $ref: '#/components/schemas/Pager'
        data:
          type: array
          items:
            $ref: '#/components/schemas/Video'
      description: ビデオページレスポンス
    VideoUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        filename:
          type: string
          minLength: 1
          maxLength: 255
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        audioActivityRatio:
          type: number
          format: double
          description: 音声アクティビティ率(%)
      description: ビデオ更新リクエスト
servers:
  - url: http://localhost:8080
    description: Development server
    variables: {}